requirements.txt
Flask==1.1.2
pillow==7.2.0
json5==0.9.5
tensorflow==2.3.0
keras=2.4.3
opencv-contrib-python==4.4.0.44
numpy==1.18.5


